Issue #34: Better handling of sql_params
The current implementation only handled one SQL parameter, and the
generated API URL was wrong.

This patch makes the code use $c->validation->every_param('sql_params')
so all parameters are taken into account.

The template for generating the opacuserjs gets updated so the generated
URI is a valid one.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>